% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{filter.date}
\alias{filter.date}
\title{filter 1: filter the Date > today or some date}
\usage{
filter.date(envir, dt = Sys.Date(), ...)
}
\description{
filter 1: filter the Date > today or some date
}

