% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WindAPIWrapper.R
\name{wind.fetchData}
\alias{wind.fetchData}
\title{fetch data}
\usage{
wind.fetchData(command, param, limit = 9, readFlag = TRUE,
  writeFlag = TRUE, filters = NULL, ...)
}
\arguments{
\item{command}{is a character with quote or without quote}

\item{param}{is a list or a string.}
}
\value{
`$Data, in data.table format
}
\description{
fetch data
}

